Saves and grabs data from Core data before downloading anything

+++ Rustar/Controllers/LocationPicker/FirstPageJSONData.swift
@@ -7,7 +7,7 @@ struct FirstPageJSONData : Decodable {
     let price: Decimal
     let old_price: Decimal
     let url : String
-    let rank : Double
+    let rank : Decimal
 }
 
 struct FirstPageJSONDataFrame : Decodable {
